refactor(state): extract applyUserFacingError to a free function
Move Model's private applyUserFacingError (which never touched `this`) into a
standalone module so it can be shared by the compile/export services about to
be extracted from Model. Behavior-identical; all six call sites updated.

Refs #59
